[REFACTOR] move id generator to helper
