+++ internal/common/helper/idgenerator_test.go
@@ -1,4 +1,4 @@
-package account
+package helper
 
 import (
 	"crypto/rand"
@@ -17,7 +17,7 @@ func Test_GenerateRandomID(t *testing.T) {
 	t.Run("success", func(t *testing.T) {
 		g := gomega.NewWithT(t)
 
-		id, err := generateRandomID()
+		id, err := GenerateRandomID()
 
 		g.Expect(err).To(gomega.BeNil())
 		g.Expect(len(id)).To(gomega.Equal(15))
@@ -26,8 +26,8 @@ func Test_GenerateRandomID(t *testing.T) {
 	t.Run("success - unique IDs", func(t *testing.T) {
 		g := gomega.NewWithT(t)
 
-		id1, err1 := generateRandomID()
-		id2, err2 := generateRandomID()
+		id1, err1 := GenerateRandomID()
+		id2, err2 := GenerateRandomID()
 
 		g.Expect(err1).To(gomega.BeNil())
 		g.Expect(err2).To(gomega.BeNil())
@@ -44,7 +44,7 @@ func Test_GenerateRandomID(t *testing.T) {
 		})
 		defer monkey.Unpatch(rand.Int)
 
-		_, err := generateRandomID()
+		_, err := GenerateRandomID()
 
 		g.Expect(err).ShouldNot(gomega.BeNil())
 	})
